[Large changes!][Registration page added and sexy!!!][Progress indicator + api check for available username][Sexy edit text boxes display material-like behavior, and also validate the format of the entries][Validation on text changing for email/password][Username on success/failure progress displays]
I kind of messed up the concise commit. There is also a checkmark icon/toolbar for submitting.

All of this was an incredible pain in my ass to do, but it's looking pretty polished. The code is not in the same state of polish, which will require perhaps some utility classes or two to tighten up the readability.

I'm still looking to make a library out of this without the nice landing page art. Just Android login + register + account details, then Server authenticate/signup/login/username check. We will see how it goes...