[UserSettings created, more extensible than meets the eye -- a little boring for now][Signup hooked up to verification, and it works to verify the user!][No tests done to try and fool the server -- oops]
Next: finish login procedure -- need login username/password on the screen preferably. Also need to wire up the facebook button to send user to registration page with pre-entered info.

Important: need user session manager singleton to manage the state of the user. Are they signed on, do they have an api_token, when does it expire, etc etc. All places in the app should be able to access the user object to pull things from it if necessary.