[Added persistence!][User loggins are stored in local preferences, so you are logged in even if you exit the app]
Next: pairing down the code, everything for login/logout is done now! Though there is a serious need for solid error handling and logging inside the application. Those will come soon. I'm going to put this into the win-java app to show ken.

One more thing, adding croutons to say hello to the user whenever the cache is fetched and they are logged in for the first time.