[Now with some major crouton action][Now tracking if the user is loaded from cache, or logs in during the session]
Next: cleanup! All the nicities are there, but they're definitely convoluted and difficult to follow. Need to clean it up and package it up.

Then, it needs to be turned into an actual library for use in other projects. I want to be able to include a login/signup screen for any project without much difficulty. Time to make a true library.