[Lots of changes -- now working on quick and dirty GPU code]
Warning: the code base is quickly getting out of control. It is difficult to understand from the outside, and poorly architected.

Once the GPU image filter method is established, the whole code base should be paired down to only the essentials, and those should be as clear as possible. No dependency injection, no hugely complicated abstractions. Create the classes we need and no more. It's too unqieldy currently.