[Working edit - start evolution, stop it, continue evolution from where you stopped locally][Need: publish to web button]
With this, I'm about ready to look at the server infrastructure for saving. There are some minor tweaks required, like adding a publish vs save button that will be less confusing for user. Otherwise, the UI for simple (Mark I) filter creation is in place with all the makings to support something more complicated.